requirements.txt
asgiref==3.3.4
Django==3.2.4
django-crispy-forms==1.11.2
django-filter==2.4.0
numpy==1.19.5
pandas==1.1.5
PyMySQL==1.0.2
python-dateutil==2.8.1
pytz==2021.1
six==1.16.0
sqlparse==0.4.1
typing-extensions==3.10.0.0
